// Client/src/app/shared/models/line-chart.ts
export class ILineChart {
  janInvoice: number;
  febInvoice: number;
  marInvoice: number;
  aprInvoice: number;
  mayInvoice: number;
  junInvoice: number;
  julInvoice: number;
  augInvoice: number;
  sepInvoice: number;
  octInvoice: number;
  novInvoice: number;
  decInvoice: number;
  janPaid: number;
  febPaid: number;
  marPaid: number;
  aprPaid: number;
  mayPaid: number;
  junPaid: number;
  julPaid: number;
  augPaid: number;
  sepPaid: number;
  octPaid: number;
  novPaid: number;
  decPaid: number;
}